+++ grakn-test/test-integration/src/test/java/ai/grakn/test/engine/postprocessing/PostProcessingTest.java
@@ -18,21 +18,29 @@
 
 package ai.grakn.test.engine.postprocessing;
 
+import ai.grakn.Grakn;
 import ai.grakn.GraknSession;
 import ai.grakn.GraknTx;
 import ai.grakn.GraknTxType;
+import ai.grakn.Keyspace;
 import ai.grakn.concept.Attribute;
 import ai.grakn.concept.AttributeType;
+import ai.grakn.concept.Concept;
 import ai.grakn.concept.ConceptId;
+import ai.grakn.concept.EntityType;
+import ai.grakn.engine.task.postprocessing.IndexPostProcessor;
+import ai.grakn.engine.task.postprocessing.InstanceCountPostProcessor;
 import ai.grakn.engine.task.postprocessing.PostProcessor;
+import ai.grakn.engine.task.postprocessing.RedisCountStorage;
+import ai.grakn.engine.task.postprocessing.RedisIndexStorage;
 import ai.grakn.exception.InvalidKBException;
 import ai.grakn.kb.log.CommitLog;
 import ai.grakn.test.rule.EngineContext;
 import ai.grakn.util.GraknTestUtil;
+import ai.grakn.util.SampleKBLoader;
 import ai.grakn.util.Schema;
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Sets;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.junit.After;
@@ -50,7 +58,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assume.assumeTrue;
 
 public class PostProcessingTest {
-    private static final ObjectMapper mapper = new ObjectMapper();
     private PostProcessor postProcessor;
     private GraknSession session;
 
@@ -58,14 +65,20 @@ public class PostProcessingTest {
     public static final EngineContext engine = EngineContext.createWithInMemoryRedis();
 
     @BeforeClass
-    public static void onlyRunOnTinker(){
+    public static void onlyRunOnTinker() {
         assumeTrue(GraknTestUtil.usingTinker());
     }
 
     @Before
-    public void setUp() throws Exception {
-        session = engine.sessionWithNewKeyspace();
-        postProcessor = PostProcessor.create(engine.config(), engine.getJedisPool(), engine.server().factory(), engine.server().lockProvider(), new MetricRegistry());
+    public void setupPostProcessor() {
+        MetricRegistry metricRegistry = new MetricRegistry();
+        RedisIndexStorage indexStorage = RedisIndexStorage.create(engine.getJedisPool(), metricRegistry);
+        IndexPostProcessor indexPostProcessor = IndexPostProcessor.create(engine.server().lockProvider(), indexStorage);
+
+        RedisCountStorage countStorage = RedisCountStorage.create(engine.getJedisPool(), metricRegistry);
+        InstanceCountPostProcessor countPostProcessor = InstanceCountPostProcessor.create(engine.config(), engine.server().factory(), engine.server().lockProvider(), metricRegistry, countStorage);
+
+        postProcessor = PostProcessor.create(indexPostProcessor, countPostProcessor);
     }
 
     @After
@@ -127,5 +140,78 @@ public class PostProcessingTest {
         assertEquals(1, tx.getAttributeType(sample).instances().count());
 
         tx.close();
+    }
+
+    @Test
+    public void whenUpdatingInstanceCounts_EnsureRedisIsUpdated() throws InterruptedException {
+        RedisCountStorage redis = engine.redis();
+        Keyspace keyspace = SampleKBLoader.randomKeyspace();
+        String entityType1 = "e1";
+        String entityType2 = "e2";
+
+        //Create Artificial configuration
+        createAndExecuteCountTask(keyspace, ConceptId.of(entityType1), 6L);
+        createAndExecuteCountTask(keyspace, ConceptId.of(entityType2), 3L);
+        // Check cache in redis has been updated
+        assertEquals(6L, redis.getCount(RedisCountStorage.getKeyNumInstances(keyspace, ConceptId.of(entityType1))));
+        assertEquals(3L, redis.getCount(RedisCountStorage.getKeyNumInstances(keyspace, ConceptId.of(entityType2))));
+
+        //Create Artificial configuration
+        createAndExecuteCountTask(keyspace, ConceptId.of(entityType1), 1L);
+        createAndExecuteCountTask(keyspace, ConceptId.of(entityType2), -1L);
+        // Check cache in redis has been updated
+        assertEquals(7L, redis.getCount(RedisCountStorage.getKeyNumInstances(keyspace, ConceptId.of(entityType1))));
+        assertEquals(2L, redis.getCount(RedisCountStorage.getKeyNumInstances(keyspace, ConceptId.of(entityType2))));
+    }
+
+    private void createAndExecuteCountTask(Keyspace keyspace, ConceptId conceptId, long count) {
+        //Create the fake commit log
+        CommitLog commitLog = CommitLog.createDefault(keyspace);
+        commitLog.instanceCount().put(conceptId, count);
+
+        //Start up the Job
+        postProcessor.index().updateIndices(commitLog);
+    }
+
+    @Test
+    public void whenShardingThresholdIsBreached_ShardTypes() {
+        Keyspace keyspace = SampleKBLoader.randomKeyspace();
+        EntityType et1;
+        EntityType et2;
+
+        //Create Simple GraknTx
+        try (GraknTx graknTx = Grakn.session(engine.uri(), keyspace).open(GraknTxType.WRITE)) {
+            et1 = graknTx.putEntityType("et1");
+            et2 = graknTx.putEntityType("et2");
+            graknTx.admin().commitSubmitNoLogs();
+        }
+
+        checkShardCount(keyspace, et1, 1);
+        checkShardCount(keyspace, et2, 1);
+
+        //Add new counts
+        createAndExecuteCountTask(keyspace, et1.getId(), 99_999L);
+        createAndExecuteCountTask(keyspace, et2.getId(), 99_999L);
+
+        checkShardCount(keyspace, et1, 1);
+        checkShardCount(keyspace, et2, 1);
+
+        //Add new counts
+        createAndExecuteCountTask(keyspace, et1.getId(), 2L);
+        createAndExecuteCountTask(keyspace, et2.getId(), 1L);
+
+        checkShardCount(keyspace, et1, 2);
+        checkShardCount(keyspace, et2, 1);
+    }
+
+    private void checkShardCount(Keyspace keyspace, Concept concept, int expectedValue) {
+        try (GraknTx graknTx = Grakn.session(engine.uri(), keyspace).open(GraknTxType.WRITE)) {
+            int shards = graknTx.admin().getTinkerTraversal().V().
+                    has(Schema.VertexProperty.ID.name(), concept.getId().getValue()).
+                    in(Schema.EdgeLabel.SHARD.getLabel()).toList().size();
+
+            assertEquals(expectedValue, shards);
+
+        }
     }
 }
